extractTtbarTruth : cleanup
Details:
no need for functor, can use simple functions (would like to make them
local though...)
The problem is not here however, we're counting b correctly; the
problem is upstream in WhTruthExtractor, where we're saving t, W and
children, but not b.
Fix that and in the meantime just look at the deltaR(jet,mu)
distributions.